% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1931_AmN_f02}
\alias{gause_1931_AmN_f02}
\title{The influence of quantity of food on the asymptotic population of Triboliurn confusum.}
\format{
A data frame with 4 rows and 5 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Figure}{Figure number in paper}
  \item{Species}{Name of Species: Tribolium confusum}
  \item{Individuals}{Number of Individuals at population equlibrium}
  \item{Treatment}{Treatments:16, 32, 64 and 128g flour per starting beetle pair}
}
}
\source{
Gause (1931) The influence of ecological factors on the size of population. The American Naturalist 65:696, 70-76
}
\usage{
gause_1931_AmN_f02
}
\description{
A dataset containing the population size of Tribolium confusum at the upper asymptote,
as a characteristic for the population beeing in equilibrium, under different quantities of food (flour).
Gause's goal was to determine the influence of ecological factors on population growth.
}
\keyword{datasets}
