% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_app_t01}
\alias{gause_1934_book_app_t01}
\title{Raw data on the abundances and volumes of Saccharomyces cerevisiae and Schizosaccharomyces kephir}
\format{
A data frame with 60 rows and 10 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Table}{Table number in paper}
  \item{Experiment}{Number of experiment: 1 or 2}
  \item{Species}{Name of Species: Saccharomyces cerevisiae and Schizosaccharomyces kephir}
  \item{Time}{Day of experiment}
  \item{Volume_Species}{Volume of yeast species in the mixture, estimated from counted cells}
  \item{Squares}{Number of Squares counted}
  \item{Individuals_Square}{Average Number of cells per Square}
  \item{Volume_Total}{Total volume of yeast}
  \item{Treatment}{Treatments: Monoculture and Mixture}
 
}
}
\source{
Gause (1934) The Struggle for Existence. Dover Publications; Reprint edition.
}
\usage{
gause_1934_book_app_t01
}
\description{
A dataset containing the abundances (number of cells) and volumes of Saccharomyces cerevisiae 
and Schizosaccharomyces kephir grown in monoculture and in mixture.
The data for two experiments with differnet time periods is reported.
}
\keyword{datasets}
