% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_app_t02}
\alias{gause_1934_book_app_t02}
\title{Alcohol production of Saccharomyces cerevisiae and Schizosaccharomyces kephir}
\format{
A data frame with 28 rows and 7 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Table}{Table number in paper}
  \item{Species}{Name of Species: Saccharomyces cerevisiae, Schizosaccharomyces kephir}
  \item{Time}{Hour of experiment}
  \item{Alcohol}{Alcohol production in percent}
  \item{Alcohol.Yeast_Volume}{Alcohol production per unit of yeast volume}
  \item{Treatment}{Treatments: Anaerobic and Aerobic}
}
}
\source{
Gause (1934) The Struggle for Existence. Dover Publications; Reprint edition.
}
\usage{
gause_1934_book_app_t02
}
\description{
A dataset containing the  alcohol production of Saccharomyces cerevisiae and 
Schizosaccharomyces kephir cultivated under anaerobic and aerobic conditions, 
in percent and per unit of yeast volume.
}
\keyword{datasets}
