% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_app_t03}
\alias{gause_1934_book_app_t03}
\title{Raw data of Paramecium caudatum and Paramecium aurelia grown in Monoculture and Mixture}
\format{
A data frame with 104 rows and 6 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Table}{Table number in paper}
  \item{Species}{Name of Species: Paramecium caudatum and Paramecium aurelia}
  \item{Time}{Day of experiment}
  \item{Individuals}{Number of Individuals}
  \item{Treatment}{Treatments: Monoculture and Mixture}
}
}
\source{
Gause (1934) The Struggle for Existence. Dover Publications; Reprint edition.
}
\usage{
gause_1934_book_app_t03
}
\description{
A dataset containing the mean abundances of Paramecium caudatum and 
Paramecium aurelia, grown in mixture and monoculture. This dataset contains the mean abundances of three
experiments. Note, that for day 20, single values are missing and only the mean is reported.
}
\keyword{datasets}
