% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_app_t04}
\alias{gause_1934_book_app_t04}
\title{Raw data of Paramecium caudatum and Paramecium aurelia grown on different media}
\format{
A data frame with 68 rows and 7 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Table}{Table number in paper}
  \item{Species}{Name of Species: Paramecium caudatum and Paramecium aurelia}
  \item{Time}{Day of experiment}
  \item{Individuals}{Number of Individuals}
  \item{Individuals_Mixture}{Number of Individuals per Species in Mixture}
  \item{Treatment}{Treatments: One Loop and Half Loop}
}
}
\source{
Gause (1934) The Struggle for Existence. Dover Publications; Reprint edition.
}
\usage{
gause_1934_book_app_t04
}
\description{
A dataset containing the abundance of Paramecium caudatum and 
Paramecium aurelia, grown in monoculture and in mixture on buffered medium with
high wild bacteria concentration ("one loop" medium) and low wild bacteria concentration 
("half loop" medium). The number of individuals is reported for monoculture and the 
number of individuals of the species in the mixture.
}
\keyword{datasets}
