% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_app_t05}
\alias{gause_1934_book_app_t05}
\title{Raw data of Stylonychia pustulata in monoculture and mixture with and Paramecium aurelia and P. caudatum}
\format{
A data frame with 575 rows and 8 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Table}{Table number in paper}
  \item{Time}{Day of experiment}
  \item{Experiment}{Number of experiment: 1,2 or 3}
  \item{Culture}{Number of culture: 1,2,3,4,5 or Mean}
  \item{Species}{Name of Species: Stylonychia pustulata, Paramecium caudatum and Paramecium aurelia}
  \item{Individuals}{Number of Individuals}
  \item{Treatment}{Treatments: Monoculture and Mixture P.aurelia and Mixture P.caudatum}
}
}
\source{
Gause (1934) The Struggle for Existence. Dover Publications; Reprint edition.
}
\usage{
gause_1934_book_app_t05
}
\description{
A dataset containing the abundances of Stylonychia pustulata grown in monoculture and mixture with P. aurelia 
and P. caudatum and the abundances of P. aurelia and P. caudautum in this mixtures on the medium of Osterhout. 
This dataset contains the raw data of three different experiments: Stylonychia pustulata grown in monoculture, in mixture with P.aurelia
and in mixture wit P.caudatum. The abundances of three to five cultures per experiment and the calculated mean abundances of the
experiments are reported. Note, that for day 20 - 25, single values are missing and only the mean is reported.
}
\keyword{datasets}
