% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_f04}
\alias{gause_1934_book_f04}
\title{Growth of Paramecium caudatum}
\format{A data frame with 8 rows and 5 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Figure}{Figure number in paper}
  \item{Species}{Name of Species: Paramecium caudatum}
  \item{Time}{Day of experiment}
  \item{Individuals}{Number of Individuals}
}}
\source{
Gause (1934) The Struggle for Existence. Dover Publications; Reprint edition.
}
\usage{
gause_1934_book_f04
}
\description{
A dataset containing the growth in abundance  of Paramecium caudatum over six days.
}
\keyword{datasets}
