% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_f09}
\alias{gause_1934_book_f09}
\title{Growth of Saccharomyces cerevisiae}
\format{
A data frame with 9 rows and 5 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Figure}{Figure number in paper}
  \item{Species}{Name of Species: Saccharomyces cerevisiae}
  \item{Time}{Hour of experiment}
  \item{Volume}{"Amount of Yeast" of Species}
}
}
\source{
Gause (1934) The Struggle for Existence. Dover Publications; Reprint edition.
}
\usage{
gause_1934_book_f09
}
\description{
A dataset containing the  growth  in  volume  of  Saccharomyces cerevisiae under anaerobic conditions.
The Volume is measured in "Amount of yeast", which refers to a standardized index, 
based on the alcohol production per Unit of Yeast.
}
\keyword{datasets}
