% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_f11}
\alias{gause_1934_book_f11}
\title{Gause Yeast Data}
\format{A data frame with 17 rows and 8 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Figure}{Figure number in paper}
  \item{Species}{Species name}
  \item{Time}{Experiment time in hours}
  \item{Alcohol_Percent}{Alcohol concentration, in percent}
  \item{Yeast_Volume}{Yeast volume, listed in papers as 'amount'}
  \item{Treatment}{Experiment replicate, under two 'somewhat different' growth medium concentrations}
}}
\source{
Huffaker (1958) Experimental Studies on Predation: Dispersion Factors and Predator-Prey Oscillations. Hilgardia 27: 343–83.
}
\usage{
gause_1934_book_f11
}
\description{
A dataset containing yeast volume and alcohol concentration
for two replicates of S. cerevisiae grown in monoculture.
}
\keyword{datasets}
