% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_f13}
\alias{gause_1934_book_f13}
\title{Growth of Saccharomyces cerevisiae and Schizosaccaromyces kephir in mixed population}
\format{
A data frame with 47 rows and 5 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Figure}{Figure number in paper}
  \item{Species}{Name of Species: Saccharomyces cerevisiae, Schizosaccaromyces kephi and Mixed Population}
  \item{Time}{Hour of experiment}
  \item{Volume}{"Amount of Yeast" of Species}
}
}
\source{
Gause (1934) The Struggle for Existence. Dover Publications; Reprint edition.
}
\usage{
gause_1934_book_f13
}
\description{
A dataset containing the the  growth  in  volume  of  Saccharomyces cerevisiae and Schizosaccaromyces  kephir cultivated 
in  the  mixed  population (two  series  of  experiments) under anaerobic conditions. Gause also measured the volume of 
the mixed population. The Volume is measured in "Amount of yeast", which refers to a standardized index, 
based on the alcohol production per Unit of Yeast.
}
\keyword{datasets}
