% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_f14}
\alias{gause_1934_book_f14}
\title{Growth of Saccharomyces cerevisiae in mixed population}
\format{
A data frame with 32 rows and 6 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Figure}{Figure number in paper}
  \item{Species}{Name of Species: Saccharomyces cerevisiae}
  \item{Time}{Hour of experiment}
  \item{Volume}{"Amount of Yeast" of Species}
  \item{Treatment}{Treatments: Monoculture and Mixture}
}
}
\source{
Gause (1934) The Struggle for Existence. Dover Publications; Reprint edition.
}
\usage{
gause_1934_book_f14
}
\description{
A dataset containing the the  growth  in  volume  of  Saccharomyces cerevisiae  cultivated 
separately  and  in  the  mixed  population (two  series  of  experiments) with Schizosaccaromyces  kephir 
under anaerobic conditions. The Volume is measured in "Amount of yeast", which refers to a standardized index, 
based on the alcohol production per Unit of Yeast.
}
\keyword{datasets}
