% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_f16}
\alias{gause_1934_book_f16}
\title{Growth of Schizosaccaromyces kephir and Saccharomyces cerevisiae in mixed population -aerobic}
\format{
A data frame with 27 rows and 6 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Figure}{Figure number in paper}
  \item{Species}{Name of Species: Schizosaccaromyces kephir, Saccharomyces cerevisiae and Total_Yeast}
  \item{Time}{Hour of experiment}
  \item{Volume}{"Amount of Yeast" of Species}
  \item{Treatment}{Treatments: Monoculture and Mixture}
}
}
\source{
Gause (1934) The Struggle for Existence. Dover Publications; Reprint edition.
}
\usage{
gause_1934_book_f16
}
\description{
A dataset containing the the  growth  in  volume  of  Schizosaccaromyces  kephir  cultivated 
separately  and  in  the  mixed  population with Saccharomyces cerevisiae under aerobic conditions.
The Volume is measured in "Amount of yeast", which refers to a standardized index, 
based on the alcohol production per Unit of Yeast.
}
\keyword{datasets}
