% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_f19}
\alias{gause_1934_book_f19}
\title{Growth of Paramecium caudatum and Stylonychia mytilis in Mixture with Wild bacteria}
\format{
A data frame with 17 rows and 6 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Figure}{Figure number in paper}
  \item{Species}{Name of Species: Paramecium caudatum and Stylonychia mytilis}
  \item{Time}{Day of experiment}
  \item{Individuals}{Number of Individuals}
  \item{Treatment}{Treatments: Monoculture and Mixture}
}
}
\source{
Gause (1934) The Struggle for Existence. Dover Publications; Reprint edition.
}
\usage{
gause_1934_book_f19
}
\description{
A dataset containing the abundance of Paramecium caudatum and 
Stylonychia mytilis, grown in monoculture and in mixture on buffered medium containing 
wild bacteria.
}
\keyword{datasets}
