% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_f21}
\alias{gause_1934_book_f21}
\title{Growth of Paramecium caudatum and Paramecium aurelia in Monoculture}
\format{A data frame with 87 rows and 6 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Figure}{Figure number in paper}
  \item{Species}{Name of Species: Paramecium caudatum and Paramecium aurelia}
  \item{Time}{Day of experiment}
  \item{Volume}{"Volume" of Species}
  \item{Individuals}{Number of Individuals}
}}
\source{
Gause (1934) The Struggle for Existence. Dover Publications; Reprint edition.
}
\usage{
gause_1934_book_f21
}
\description{
A dataset containing the abundances and the volume of Paramecium caudatum and 
Paramecium aurelia, to determine the differences in reaching the saturating population 
regarding Volume and Number of individuals.
"Volume" refers to a standardized index, meant to make the abundances
of species comparable based on their relative sizes.
}
\keyword{datasets}
