% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_f22}
\alias{gause_1934_book_f22}
\title{Paramecium competition experiment}
\format{
A data frame with 72 rows and 4 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Figure}{Figure number in paper}
  \item{Day}{Day of experiment}
  \item{Species1}{Name of Species 1}
  \item{Volume_Species1}{Volume of Paramecium caudatum}
  \item{Species2}{Name of Species 2}
  \item{Volume_Species2}{Volume of Paramecium aurelia}
  \item{Treatment}{Treatments: Pa and Pc moncultures, or mixture}
}
}
\source{
Gause (1934) The Struggle for Existence. Dover Publications; Reprint edition.
}
\usage{
gause_1934_book_f22
}
\description{
A dataset containing the abundances of two Paramecium species
grown in monoculture and mixture. Note, is for the same experiment
as gause_1934_science_f02_03, except that data were digitized separately,
and therefore have small variations. These might be useful for
estimating observation error in the data digitization process.
"Volume" refers to a standardized index, meant to make the abundances
of species comparable based on their relative sizes.
}
\keyword{datasets}
