% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_f23}
\alias{gause_1934_book_f23}
\title{Growth of Paramecium caudatum and Paramecium aurelia with wild bacteria}
\format{A data frame with 61 rows and 6 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Figure}{Figure number in paper}
  \item{Species}{Name of Species: Paramecium caudatum and Paramecium aurelia}
  \item{Time}{Day of experiment}
  \item{Volume}{"Volume" of Species}
  \item{Treatment}{Treatments: One Loop and Half Loop}
}}
\source{
Gause (1934) The Struggle for Existence. Dover Publications; Reprint edition.
}
\usage{
gause_1934_book_f23
}
\description{
A dataset containing the volume of Paramecium caudatum and 
Paramecium aurelia, grown in monoculture on buffered medium with two different 
wild bacteria concentrations ("one" and "half loop").
"Volume" refers to a standardized index, meant to make the abundances
of species comparable based on their relative sizes.
}
\keyword{datasets}
