% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_f25}
\alias{gause_1934_book_f25}
\title{Growth of Paramecium caudatum and Paramecium aurelia in Mixture on One Loop Medium}
\format{
A data frame with 57 rows and 6 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Figure}{Figure number in paper}
  \item{Species}{Name of Species: Paramecium caudatum and Paramecium aurelia}
  \item{Time}{Day of experiment}
  \item{Volume}{"Volume" of Species}
  \item{Treatment}{Treatments: Monoculture and Mixture}
}
}
\source{
Gause (1934) The Struggle for Existence. Dover Publications; Reprint edition.
}
\usage{
gause_1934_book_f25
}
\description{
A dataset containing the volume of Paramecium caudatum and 
Paramecium aurelia, grown in monoculture and in mixture on buffered medium with
high wild bacteria concentration ("one loop" medium).
"Volume" refers to a standardized index, meant to make the abundances
of species comparable based on their relative sizes.
}
\keyword{datasets}
