% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_f26}
\alias{gause_1934_book_f26}
\title{Growth of Stylorzychia  pustulata in Monoculture and in Mixture}
\format{
A data frame with 104 rows and 6 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Figure}{Figure number in paper}
  \item{Species}{Name of Species:Paramecium caudatum, Paramecium aurelia and Stylorzychia  pustulata}
  \item{Time}{Day of experiment}
  \item{Individuals}{Number of Individuals}
  \item{Treatment}{Treatments: Monoculture and Mixture(with Species)}
}
}
\source{
Gause (1934) The Struggle for Existence. Dover Publications; Reprint edition.
}
\usage{
gause_1934_book_f26
}
\description{
A dataset containing the abundances of  Stylorzychia  pustulata cultivated  separately,  
and  in  the  mixed  populations  with  Paramecium caudatum and Paramecium aurelia and the 
abundances of P. caudatum and P. aurelia grown in mixture with S. pustulata.
}
\keyword{datasets}
