% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_f28}
\alias{gause_1934_book_f28}
\title{Elementary interaction between Didinium nasutum and Paramecium caudatum}
\format{
A data frame with 12 rows and 6 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Figure}{Figure number in paper}
  \item{Species}{Name of Species: Paramecium caudatum and Didinium nasutum}
  \item{Time}{Day of experiment}
  \item{Individuals}{Number of Individuals} 
  \item{Treatment}{Treatments: Monoculture and Mixture}
}
}
\source{
Gause (1934) The Struggle for Existence. Dover Publications; Reprint edition.
}
\usage{
gause_1934_book_f28
}
\description{
A dataset containing the abundances of Paramecium caudatum and Didinium nasutum
grown in mixture. Didinium was introduced at day two. 
"Individuals" refers to the number of individuals pro 0.5 c.c..
}
\keyword{datasets}
