% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_f29}
\alias{gause_1934_book_f29}
\title{Paramecium/Didinium predator-prey experiment}
\format{
A data frame with 62 rows and 7 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Figure}{Figure number in paper}
  \item{Species}{Name of Species: Didinium nasutum and Paramecium caudatum}
  \item{Time}{Day of experiment}
  \item{Volume}{Volume of Didinium}
  \item{Individuals}{Number of Individuals}
  \item{Treatment}{Treatments: D. nasutum introduced after 0, 24, 36 and 48 hrs}
}
}
\source{
Gause (1934) The Struggle for Existence. Dover Publications; Reprint edition.
}
\usage{
gause_1934_book_f29
}
\description{
A dataset containing the abundances of Paramecium caudatum and Didinium nasutum
grown in mixture. Didinium was introduced at different days. 
"Volume" refers to a standardized index, meant to make the abundances
of species comparable based on their relative sizes.
}
\keyword{datasets}
