% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_f31}
\alias{gause_1934_book_f31}
\title{The interaction between Didinium nasutum and Paramecium caudatum on oat medium}
\format{A data frame with 12 rows and 6 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Figure}{Figure number in paper}
  \item{Time}{Day of experiment}
  \item{Species}{Name of Species: Didinium nasutum and Paramecium caudatum}
  \item{Individuals}{Number of Individuals}
  \item{Treatment}{Treatment: Oat medium}
}}
\source{
Gause (1934) The Struggle for Existence. Dover Publications; Reprint edition.
}
\usage{
gause_1934_book_f31
}
\description{
A dataset containing the abundances of Didinium nasutum and Paramecium caudatum grown in mixture on oat medium with sediment.
}
\keyword{datasets}
