% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1934_book_f39.1}
\alias{gause_1934_book_f39.1}
\title{The interaction between Paramecium bursaria and Schizosaccharomyces pombe}
\format{
A data frame with 36 rows and 5 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Figure}{Figure number in paper}
  \item{Time}{Day of experiment}
  \item{Species}{Name of Species: Paramecium bursaria and Schizosaccharomyces pombe}
  \item{Individuals}{Number of Individuals}
}
}
\source{
Gause (1934) The Struggle for Existence. Dover Publications; Reprint edition.
}
\usage{
gause_1934_book_f39.1
}
\description{
A dataset containing the abundances of Paramecium bursaria and Schizosaccharomyces pombe grown in mixture.
}
\keyword{datasets}
