% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1936_AnE_f01}
\alias{gause_1936_AnE_f01}
\title{Interaction between predators (Cheyletus eruditus) and prey (Aleuiroglyphus agilis) on millet, wheat flour and a mixture of these substances.}
\format{A data frame with 34 rows and 6 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Figure}{Figure number in paper}
  \item{Time}{Day of experiment}
  \item{Species}{Name of Species: Cheyletus eruditus and Aleuiroglyphus agilis}
  \item{Individuals}{Number of Individuals per 0.2 g}
  \item{Treatment}{Treatments: Wheat, Millet, Wheat+Millet}
}}
\source{
Gause (1936) Further Studies of Interaction between Predators and Prey. Journal of Animal Ecology, vol. 5, pp. 1-18.
}
\usage{
gause_1936_AnE_f01
}
\description{
A dataset containing the abundance of Cheyletus eruditus and Aleuiroglyphus agilis,
as apredator-prey system under different food conditions for the prey (wheat and millet).
Gause?s goal was to determine the influence of ecological factors on predator-prey dynamics.
The number of individuals is reported as Individuals per 0.2 g prey food.
}
\keyword{datasets}
