% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1936_AnE_f03.3b}
\alias{gause_1936_AnE_f03.3b}
\title{Interaction  between  predators  (Cheyletus eruditus) and  prey  (Aleuroglyphus agilis) with artifical immigration}
\format{
A data frame with 26 rows and 6 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Figure}{Figure number in paper}
  \item{Time}{Day of experiment}
  \item{Species}{Name of Species: Cheyletus eruditus and Aleuiroglyphus agilis}
  \item{Individuals}{Number of Individuals per 0.2 g}
  \item{Treatment}{Treatment: Immigration}
}
}
\source{
Gause (1936) Further Studies of Interaction between Predators and Prey. Journal of Animal Ecology, vol. 5, pp. 1-18.
}
\usage{
gause_1936_AnE_f03.3b
}
\description{
A dataset containing the abundance of Cheyletus eruditus and Aleuiroglyphus agilis,
as apredator-prey system with  an artificial immigration  of  predator and prey every 8th day.
The number of individuals is reported as Individuals per 0.2 g prey food. Wheat flour was used as food.
}
\keyword{datasets}
