% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1936_AnE_t02}
\alias{gause_1936_AnE_t02}
\title{Interaction  between  predators  (Cheyletus eruditus) and  prey  (Aleuroglyphus agilis)-Raw data}
\format{
A data frame with 191 rows and 12 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Table}{Table number in paper}
  \item{Time}{Day of experiment}
  \item{Species}{Name of Species: Cheyletus eruditus and Aleuiroglyphus agilis}
  \item{Total_Individuals}{Number of all Individuals of Species per 0.2 g}
  \item{Female}{Number of female individuals per 0.2 g}
  \item{Male}{Number of male individuals per 0.2 g}
  \item{Imago}{Number of adult Individuals per 0.2 g}
  \item{Hexapod}{Number of hexapod larvae stage Individuals per 0.2 g}
  \item{Octapod}{Number of octapod larvae stage Individuals per 0.2 g}
  \item{Initial_Fraction}{Initial Number of prey/predator}
  \item{Treatment}{Treatments: Wheat and Semoletta}
}
}
\source{
Gause (1936) Further Studies of Interaction between Predators and Prey. Journal of Animal Ecology, vol. 5, pp. 1-18.
}
\usage{
gause_1936_AnE_t02
}
\description{
A dataset containing the abundance of Cheyletus eruditus and Aleuiroglyphus agilis,
as a predator-prey system on semoletta and wheat flour.  A fraction at the beginning of  
each experiment shows the initial  relation between predators and prey;  e.g. 10/5 means 10 prey 
and 5 predators (reported in variable "Treatment"). Dataset includes age structured population abundances.
The number of individuals is reported as Individuals per 0.2 g prey food. Wheat flour was used as food.
}
\keyword{datasets}
