% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gause_1936_AnE_t03}
\alias{gause_1936_AnE_t03}
\title{Interaction  between  predators  (Paramecium bursaria) and  prey  (Saccharomyces exiguus)}
\format{
A data frame with 266 rows and 6 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Table}{Table number in paper}
  \item{Experiment}{Experimental replicate number}
  \item{Time}{Day of experiment}
  \item{Species}{Name of Species: Cheyletus eruditus and Aleuiroglyphus agilis}
  \item{Individuals}{Number of all Individuals of Species per 0.2 g}
}
}
\source{
Gause (1936) Further Studies of Interaction between Predators and Prey. Journal of Animal Ecology, vol. 5, pp. 1-18.
}
\usage{
gause_1936_AnE_t03
}
\description{
A dataset containing the abundance of Paramecium bursaria and Saccharomyces exiguus,
as a predator-prey system.  The primary difference among experimental replicates
is the initial abundance of the two species.
The number of individuals is reported as Individuals per 0.5 cm3 for Paramecium,
and as number of individuals per 0.1 cm3 for Saccharomyces.
}
\keyword{datasets}
