% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mclaren_1994_f03}
\alias{mclaren_1994_f03}
\title{Wolf, Moose, and Fir dynamics from Isle Royale}
\format{
A data frame with 140 rows and 7 variables:
\describe{
  \item{Paper}{Paper from which data are drawn}
  \item{Figure}{Figure number in paper}
  \item{year}{Year of measurements}
  \item{Species}{Species name}
  \item{width}{Width of tree rings}
  \item{individuals}{Number of wolf or moose individuals}
  \item{AET..mm.}{AET water availablility index}
}
}
\source{
McLaren & Peterson (1994) Wolves, Moose, and Tree Rings on Isle Royale. Science 266:1555-1558.
}
\usage{
mclaren_1994_f03
}
\description{
A dataset containing the abundances of wolves, moose, and fir trees
from the Isle Royale study of McLaren et al.
}
\keyword{datasets}
